refactor: find-id 및 find-password 컴포넌트/에셋 폴더를 auth 도메인 하위로 이동
